template-ai/pyproject.toml
[tool.poetry]
name = "template-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
langchain = "^0.3.3"
langchain-openai = "^0.2.2"
python-dotenv = "^1.0.1"
docxtpl = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
